requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
biopython==1.73
boddle==0.2.8
bottle==0.12.16
cffi==1.12.3
importlib-metadata==0.17
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
numpy==1.16.3
packaging==19.0
pandas==0.24.2
pathlib2==2.3.3
pluggy==0.12.0
py==1.8.0
pycparser==2.19
PyMySQL==0.9.3
pyparsing==2.4.0
pytest==4.6.2
python-dateutil==2.8.0
pytz==2019.1
PyX==0.14.1
roman==3.2
rpy2==3.0.4
simplegeneric==0.8.1
six==1.12.0
SQLAlchemy==1.3.3
tzlocal==1.5.1
wcwidth==0.1.7
weblogo==3.6.0
zipp==0.5.1
